added async axios-function in searchbar-component to get autocomplete-options from external api
